Add WikidataIdsDecorator::Links class
Having the top level `WikidataIdsDecorator` constant be a class causes
problems with Bundler because the `version.rb` file has to require
scraped so that it can inherit from the base `Decorator` class, but at
the point where it processes that require it's still loading in the
gemspec, so hasn't properly initialized the bundle.

You can see an example of the error this was causing in this job:

https://travis-ci.org/everypolitician/wikidata_ids_decorator/jobs/213815867

Making the top-level constant a dependency-less module isolates us from
any Bundler/RubyGems surprises.